Split base64 into three utils/ files. Add more test cases.
