+++ src/utils/pem.js
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {base64DecodeToBytes} from './base64';
+
+
+/**
+ * Converts a text in PEM format into a binary array buffer.
+ * @param {string} pem
+ * @return {!Uint8Array}
+ * @visibleForTesting
+ */
+export function pemToBytes(pem) {
+  pem = pem.trim();
+
+  // Remove pem prefix, e.g. "----BEGIN PUBLIC KEY----".
+  pem = pem.replace(/^\-+BEGIN[^-]*\-+/, '');
+
+  // Remove pem suffix, e.g. "----END PUBLIC KEY----".
+  pem = pem.replace(/\-+END[^-]*\-+$/, '');
+
+  // Remove line breaks.
+  pem = pem.replace(/[\r\n]/g, '').trim();
+
+  return base64DecodeToBytes(pem);
+}
